query ROOT for the active overlay list
The SYSROOT is for use while building (src_*) while the ROOT is for use
when merging packages (pkg_*).  Have the user logic utilize ROOT to find
the overlay logic so that a single binary package built against a SYSROOT
can be merged against multiple ROOTs.

Otherwise, when we try sharing binpkgs (like stumpy with stumpy_moblab),
the package will remember the original SYSROOT and try to re-use that.
But there's no guarantee that the SYSROOT still exists.

BUG=chromium:370487
TEST=`cbuildbot stumpy_moblab-release sandybridge-ivybridge-full-group` passes

Change-Id: Id69cecd60f597b568238766c74dfd502244120c2
Reviewed-on: https://chromium-review.googlesource.com/200252
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>