Remove race in enewuser, enewgroup
We had an obvious TOCTOU problem in enewuser and enewgroup.
The code would check to see if a user or group was already
created early in these functions, then do a bunch of work
and check again. If the account was created in the meantime,
it'd barf on the second check. This check is redundant anyway
so remove it.

BUG=chromium:378270
TEST=remove accounts for a package and emerge it again
TEST=release trybot

Change-Id: If522df57468cd79466c27945b2240d9bc11ca503
Reviewed-on: https://chromium-review.googlesource.com/201831
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>