Add a poison-pill field to deprecated accounts
CrOS has some users and groups that are dead, and the UID/GIDs should not be re-used.
Enshrine this knowledge by adding a 'defunct' field to the
user and group specifications that should no longer be used.
The build will die if someone attempts to install such a user.

BUG=chromium:376518
TEST=create an ebuild that tries to 'enewuser tor' and emerge it

Change-Id: I9651824fa7489a3eeae0d8495acdee9a5caf5f15
Reviewed-on: https://chromium-review.googlesource.com/202180
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>