[moblab] add user apache to group moblab.
On moblab, the web frontend will soon have a setup page that will
edit the shadow_config of autotest and will also be able to upload
the boto file to use for testing.

In order to allow these processes to work user apache needs to be
able to edit some files owned by user moblab. To address this apache
is being added to the moblab group.

BUG=chromium:396694
TEST=apache is in group moblab.

Change-Id: I5f14300fcf80609e7f287c0d484eb1cdef4939ec
Reviewed-on: https://chromium-review.googlesource.com/212323
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>